# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tsllm",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/runs": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Runs",
        "operationId": "runs_api_runs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunSummary"
                  },
                  "type": "array",
                  "title": "Response Runs Api Runs Get"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Submit",
        "operationId": "submit_api_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSubmittedOut"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Run",
        "operationId": "run_api_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetailOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/runs/{run_id}/cancel": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Cancel",
        "operationId": "cancel_api_runs__run_id__cancel_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatus"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/runs/{run_id}/events": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Events",
        "description": "SSE of events.jsonl. Event ids are line numbers; the stream starts after the given id.",
        "operationId": "events_api_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "last-event-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/event-stream": {}
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/runs/{run_id}/predictions": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Predictions",
        "description": "Without origin_time: the thinned origin list. With origin_time: the rows of that origin.",
        "operationId": "predictions_api_runs__run_id__predictions_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "split",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "val",
                "cal",
                "test"
              ],
              "type": "string",
              "title": "Split"
            }
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Channel"
            }
          },
          {
            "name": "origin_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Origin Time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "default": 500,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionsOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BackboneConfig": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name",
            "description": "Backbone registry name."
          },
          "checkpoint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkpoint",
            "description": "Hugging Face repository or local checkpoint directory."
          },
          "revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision",
            "description": "Checkpoint revision to resolve."
          },
          "device": {
            "type": "string",
            "enum": [
              "cuda",
              "cpu"
            ],
            "title": "Device",
            "description": "Execution device.",
            "default": "cuda"
          },
          "dtype": {
            "type": "string",
            "enum": [
              "fp32",
              "bf16"
            ],
            "title": "Dtype",
            "description": "CUDA numeric type.",
            "default": "bf16"
          },
          "options": {
            "additionalProperties": true,
            "type": "object",
            "title": "Options",
            "description": "Adapter-specific options."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "BackboneConfig"
      },
...
      "ClassifyTaskConfig": {
        "properties": {
          "type": {
            "type": "string",
            "const": "classify",
            "title": "Type",
            "description": "Task type.",
            "default": "classify"
          },
          "context_length": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Context Length",
            "description": "Context length in grid steps."
          },
          "eval_splits": {
            "items": {
              "type": "string",
              "enum": [
                "val",
                "cal",
                "test"
              ]
            },
            "type": "array",
            "minItems": 1,
            "title": "Eval Splits",
            "description": "Evaluated splits."
          },
          "train_origin_stride": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Train Origin Stride",
            "description": "Training origin stride.",
            "default": 1
          },
          "eval_origin_stride": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Eval Origin Stride",
            "description": "Evaluation origin stride.",
            "default": 1
          },
          "max_train_origins": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Max Train Origins",
            "description": "Evenly spaced training origin limit.",
            "default": 20000
          },
          "max_eval_origins": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Eval Origins",
            "description": "Evenly spaced evaluation origin limit; null is no limit."
          },
          "label": {
            "$ref": "#/components/schemas/FutureEventRule",
            "description": "Future event label rule."
          },
          "head": {
            "$ref": "#/components/schemas/HeadConfig",
            "description": "Classifier head."
          },
          "embed_batch_size": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Embed Batch Size",
            "description": "Representation batch size.",
            "default": 64
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "context_length",
          "label"
        ],
        "title": "ClassifyTaskConfig"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "detail": {
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "ErrorOut": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorOut"
      },
      "FinetuneConfig": {
        "properties": {
          "learning_rate": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Learning Rate",
            "description": "Optimizer learning rate.",
            "default": 1e-05
          },
          "num_steps": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Num Steps",
            "description": "Number of optimizer steps.",
            "default": 1000
          },
          "batch_size": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Batch Size",
            "description": "Training batch size.",
            "default": 64
          },
          "eval_every": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Eval Every",
            "description": "Validation interval in steps.",
            "default": 100
          },
          "grad_clip": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Grad Clip",
            "description": "Maximum gradient norm.",
            "default": 1.0
          },
          "lora": {
            "$ref": "#/components/schemas/LoraSpec",
            "description": "Low-rank adapter settings."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "FinetuneConfig"
      },
      "ForecastTaskConfig": {
        "properties": {
          "type": {
            "type": "string",
            "const": "forecast",
            "title": "Type",
            "description": "Task type.",
            "default": "forecast"
          },
          "context_length": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Context Length",
            "description": "Context length in grid steps."
          },
          "horizon": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Horizon",
            "description": "Forecast horizon in grid steps."
          },
          "eval_leads": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Eval Leads",
            "description": "Lead steps (1-based) that get separate metrics."
          },
          "eval_splits": {
            "items": {
              "type": "string",
              "enum": [
                "val",
                "cal",
                "test"
              ]
            },
            "type": "array",
            "minItems": 1,
            "title": "Eval Splits",
            "description": "Evaluated splits."
          },
          "train_origin_stride": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Train Origin Stride",
            "description": "Training origin stride. Adapters sample forecast training windows from fit segments, so forecast runs do not use this value.",
            "default": 1
          },
          "eval_origin_stride": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Eval Origin Stride",
            "description": "Evaluation origin stride.",
            "default": 1
          },
          "max_eval_origins": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Eval Origins",
            "description": "Evenly spaced evaluation origin limit; null is no limit."
          },
          "eval_batch_size": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Eval Batch Size",
            "description": "Forecast batch size.",
            "default": 64
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "context_length",
          "horizon"
        ],
        "title": "ForecastTaskConfig"
      },
...
      "PredictionsOut": {
        "properties": {
          "split": {
            "type": "string",
            "title": "Split"
          },
          "origin_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Time",
            "description": "Requested origin; null for the origin list."
          },
          "origins": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "date-time"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origins",
            "description": "Thinned origin list without origin_time."
          },
          "total_origins": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Origins",
            "description": "Origin count before thinning."
          },
          "rows": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows",
            "description": "Prediction rows of origin_time."
          }
        },
        "type": "object",
        "required": [
          "split",
          "origin_time",
          "origins",
          "total_origins",
          "rows"
        ],
        "title": "PredictionsOut"
      },
      "RunConfig": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name",
            "description": "Run name."
          },
          "dataset": {
            "type": "string",
            "minLength": 1,
            "title": "Dataset",
            "description": "Dataset id."
          },
          "task": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ForecastTaskConfig"
              },
              {
                "$ref": "#/components/schemas/ClassifyTaskConfig"
              }
            ],
            "title": "Task",
            "description": "Task configuration.",
            "discriminator": {
              "propertyName": "type",
              "mapping": {
                "classify": "#/components/schemas/ClassifyTaskConfig",
                "forecast": "#/components/schemas/ForecastTaskConfig"
              }
            }
          },
          "backbone": {
            "$ref": "#/components/schemas/BackboneConfig",
            "description": "Backbone configuration."
          },
          "mode": {
            "type": "string",
            "enum": [
              "zero_shot",
              "lora",
              "head",
              "full"
            ],
            "title": "Mode",
            "description": "Training mode; classification requires head."
          },
          "finetune": {
            "$ref": "#/components/schemas/FinetuneConfig",
            "description": "Fine-tuning settings; zero_shot ignores them."
          },
          "seed": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Seed",
            "description": "Random seed.",
            "default": 0
          },
          "notes": {
            "type": "string",
            "title": "Notes",
            "description": "Free-text notes.",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "dataset",
          "task",
          "backbone",
          "mode"
        ],
        "title": "RunConfig"
      },
      "RunDetailOut": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "job": {
            "$ref": "#/components/schemas/JobSpec"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "env": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Env"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "job",
          "status",
          "env"
        ],
        "title": "RunDetailOut"
      },
      "RunStatus": {
        "properties": {
          "state": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "cancelled",
              "interrupted"
            ],
            "title": "State"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "pid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pid"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "state",
          "created_at"
        ],
        "title": "RunStatus"
      },
      "RunSubmittedOut": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status"
        ],
        "title": "RunSubmittedOut"
      },
      "RunSummary": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "enum": [
              "experiment",
              "ingest"
            ],
            "title": "Kind"
          },
          "dataset": {
            "type": "string",
            "title": "Dataset"
          },
          "task": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task"
          },
          "backbone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backbone"
          },
          "mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "state": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "cancelled",
              "interrupted"
            ],
            "title": "State"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "primary_metric": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Metric"
          },
          "primary_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Value"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "name",
          "kind",
          "dataset",
          "task",
          "backbone",
          "mode",
          "state",
          "created_at",
          "finished_at",
          "primary_metric",
          "primary_value"
        ],
        "title": "RunSummary"
      },
...
    }
  }
}